Move remaining `DocumentSyncWorker` code into main app
- Move document classes into app as `PublishingApiDocument::*`
- Get rid of `DocumentSyncWorker` module
- Move gems from separate group back into main group
- Fix unqualified `logger` statement in Rake task
