+++ spec/models/publishing_api_document/unpublish_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe DocumentSyncWorker::Document::Unpublish do
+RSpec.describe PublishingApiDocument::Unpublish do
   subject(:document) { described_class.new(document_hash) }
 
   let(:repository) do
